+++ src/Lazy/SchemaSqlBuilder/PgsqlDriver.php
@@ -85,13 +85,15 @@ class PgsqlDriver
         return $sql;
     }
 
-    public function build(SchemaDeclare $schema)
+    public function build(SchemaDeclare $schema, $rebuild = false )
     {
         $sqls = array();
 
-        $sqls[] = 'DROP TABLE IF EXISTS ' 
-            . $this->driver->getQuoteTableName( $schema->getTable() )
-            . ' CASCADE';
+        if( $rebuild ) {
+            $sqls[] = 'DROP TABLE IF EXISTS ' 
+                . $this->driver->getQuoteTableName( $schema->getTable() )
+                . ' CASCADE';
+        }
 
 
         $createSql = 'CREATE TABLE ' . $this->driver->getQuoteTableName($schema->getTable()) . "( \n";
